Issue #229: Part of inventory management...
In Home/Category page, products which have a quantity of 0, but their status is available, is shown as available, but, when the user clicks on the AddToCart button, they will get an error of "Product is NOT available"!
Solution: In the SingleProduct partial view, both Status and Quantity is checked for availability (as opposed to just its status).