-Bug Fix: OrderService - QuantityRollBack problem in the fetch query of the order...
Apparently, sometimes you absolutely MUST end your postgresql queries with a semicolon, or else it won't work, even if you only have one statement. On the other hand, there are times that if you put a semicolon at the end of your query, you'd be in big trouble, as it won't even run the damn thing! It throws an exception! For the exact same thing that it forced you to do in a very similar situation! Damn their oily hides! With their semantic rules and whatnot - which don't make any sense!