-Add the project EskadDataBringer to MahtaKala solution, which is responsible for copying the Merchandise data from Eskad's SQL server database into Mahta's postgres database server.
p.s. Until now, this was a stand-alone project, written by a friend of folani's, and, only the compiled project was available here, and not the code itself, so, it was unmodifiable! Now, the connection string for Eskad's data server has had a change, and, since we couldn't change anything in that project, we had to start over! Now, this is a revised version of that project, after going through a decompile process.